token-server/requirements.txt
livekit-api~=0.8
aiohttp~=3.9
